pyproject.toml
[project]
name = "evohive"
version = "3.0.0"
description = "进化式集体智能引擎 — 让1000个AI互相厮杀，活下来的就是最好的方案"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "typer>=0.9.0",
    "litellm @ https://github.com/BerriAI/litellm/archive/refs/tags/v1.82.3-stable.patch.2.tar.gz",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "httpx>=0.24.0",
]

[project.scripts]
evohive = "evohive.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
full = [
    "httpx>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["evohive*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
